drm/gma500: Use "Broadcast RGB" property from DRM core
gma500 supports limited-range RGB output only for DisplayPort and defines a
"Broadcast RGB" property similar to the one we previously had in i915.
However, it doesn't provide proper automatic selection of the default
quantization range as specified by CTA-861. This commit removes the
property defined by gma500 in favor of the one recently introduced in the
DRM core. As a side effect, this also changes the behavior to be compliant
with CTA-861.

This also removes some dead code related to RGB quantization range handling
for SDVO outputs that was never completed.

Signed-off-by: Yussuf Khalil <dev@pp3345.net>